CMake: Disable FPE trapping and replace continue()
Floating point exception (FPE) trapping is not compatible with recent
LAPACK, since it create and handles exceptions itself. Traps enabled
with compiler flags are set on the main program, so it's not possible
to compile LAPACK without traps and use it in another program. Therefore
we have to disable the traps for now, or all "Garble" builds will fail.

The "continue()" function is not available in CMake 2.8.11, so use
a workaround