Replace xeigen with dgeev wrapper
xeigen was an old subroutine (eigen) from ieeecss/cascade
(http://www.netlib.org/ieeecss/cascade/) with some EISPACK dependencies.
Replace it with a call to the corresponding LAPACK routine (dgeev).
This required importing dgeev and dependencies in the internal library,
since dgeev was not used before. The old EISPACK dependencies can be
either removed or moved to casvb_util (only code that uses them, where
other EISPACK routines can be found).

As a side effect, modification of delayed_util is now simplified
with a helper script to write include files.