CMake: Fix external grid_it
There is a little mess because the external code is not separated
in "utility" and "program" directories, and some routines in the
molcas library refer to the external code, which should then be
compiled in the library too. Implemented a workaround where some
programs may have their code (except "main.f") included in the
molcas library.