pymolcas: Fix problem with executable files
Usually pymolcas tries to decide if a command line argument is an input
file or a sbin script by searching the name in sbin and testing for
execution permissions. This however fails if the argument is an absolute
path and the input file is executable. It is better if sbin is only
searched if the argument has no path (relative or absolute) at all.