CI: split test jobs
Shared runners on gitlab.com now time out after 3h. Unfortunately,
some jobs can take longer than that when the runner is going slow.
To avoid it, split the testing in different subsets, which required
implementing a --fromfile option to verify, in order to read the
list of tests from a file.