Increase basis label size
For high angular momentum (which with RICD start to occur earlier),
the m number can be 10 or larger, so it doesn't fit in one character.
This commit uses two characters for the m number (or m_x, m_y and m_z),
which requires increasing the label length and affects a few places.
Hopefully it doesn't introduce new bugs, but probably the RunFile and
JobIph files will be incompatible with earlier versions.