Fix bug with CASPT2+HDF5
This should have been detected before, but in most systems the
file-id for the reference wfn file is reused for the pt2 wfn.
This didn't happen with intel-arch, revealing that the program
was trying to read from a closed file.

Also disable tinker for the intel test, since we can't make it
work with the docker image.