Fix test 252
This test was actually affected by a bug. The frequencies were
calculated using atomic masses inferred from the effective nuclear
charges. Since ECP are used, the nuclear charges here are 7, and
the program was computing frequencies as if the molecule were N2.
Now with the use of actual masses the frequencies should be "correct".