Correctly handle properties of different "type"
When the smmat subroutine picks the property from the spin-free list,
it needs to know the type (singlet/triplet). Instead of passing an
additional variable, now we pass only the index in the SO list, and
it will figure out the right spin-free property (for the special cases
of SPIN and TMOM0, the "spin component" is also passed).

Also make sure the "singlet" MLTPL 1 is used for the spin-free properties.