Write subset of TDMs to HDF5 file
This supports the SUBSET keyword together with TRD1. However,
if spin-orbit is enabled, the final SO densities will not be
obtainable from the stored TDMs, since some required TDMs may
not have been saved. The output will still be OK, but the HDF5
file will be "incomplete" in a way.