CMake: Separate preprocessor flags
Use separate flag to force preprocessing of the Fortran files.
This would normally be done by using uppercase file extensions,
but for whatever reason we don't use that. When using the Ninja
generator, compilation is split into preprocessing and compilation
stage, and the compilation will give spurious warnings (with
gfortran) if an already-preprocessed is preprocessed (i.e. by using
-cpp -fpreprocessed). To avoid this, we must be able to disable
the preprocessing flag when using Ninja (it will add it automatically
where it needs it).