Natural Transition Orbital Calculation
The natural transition orbital (NTO) calculation can be performed in the RASSI module.
Compared to the one performed in the WFA module, this NTO calculation can calculate NTOs for states in different symmetries
It also saves NTO files in the Scratch directory in the format of other OpenMolcas orbital files, for example, .ScfOrb.
The NTO files can be converted to .lus files via GRID_IT module.