Test the availability of numpy
Unfortunately, due to a recently introduced regression in gfortran
it is not possible to update the docker image to install numpy
(it would also upgrade gfortran and break every BIGOT compilation).
Instead, introduce a hack to check the availability of numpy before
running the dynamixtools test, and skip it not available.