RASSCF/ALASKA: Fix SS gradients after SA
A SA gradient calculation computes and stores a "variational" density that's
used for the gradients. A subsequent SS gradient calculation would use the
leftover variational density instead of the correct SS density, resulting in
wrong gradients (or a crash with RICD). This fixes it by wiping out the
variational density at every RASSCF calculation.